reward system - points, badges, leaderboard

+++ src/hooks/useTasks.tsx
@@ -1,10 +1,9 @@
 import { useState, useEffect } from "react";
-import { Task, weekTaskState } from "../atoms/tasksAtom";
+import { Task } from "../atoms/tasksAtom";
 import {
   addDoc,
   collection,
   query,
-  getDocs,
   where,
   updateDoc,
   doc,
@@ -13,12 +12,15 @@ import {
 } from "firebase/firestore";
 import { firestore } from "../firebase/clientApp";
 import { User } from "firebase/auth";
-import { useRecoilState } from "recoil";
+import { useToast } from "@chakra-ui/react";
+import useUserPoints from "./useUserPoints";
 
-export const useTasks = (date: string, user: User) => {
+const useTasks = (date: string, user: User) => {
   const [tasks, setTasks] = useState<Task[]>([]);
-  const [currentWeekTasksState, setWeekTasksState] =
-    useRecoilState(weekTaskState);
+  const { userPoints, computePointsForTask, updatePoints } =
+    useUserPoints(user);
+
+  const toast = useToast();
 
   const handleAddTask = async (
     task: string,
@@ -49,14 +51,35 @@ export const useTasks = (date: string, user: User) => {
   };
 
   const handleCompleteTask = async (id: string) => {
+    const taskToUpdate = tasks.find((task) => task.id === id);
+
+    if (!taskToUpdate) {
+      console.error(`Task with id ${id} not found.`);
+      return;
+    }
+
+    // Clone the task and toggle its completed status
+    const updatedTask = { ...taskToUpdate, completed: !taskToUpdate.completed };
+
     const updatedTasks = tasks.map((task) =>
-      task.id === id ? { ...task, completed: !task.completed } : task
+      task.id === id ? updatedTask : task
     );
+
     setTasks(updatedTasks);
+
+    // Calculate points using the computePointsForTask function
+    const newPoints = computePointsForTask(
+      taskToUpdate,
+      updatedTask,
+      userPoints
+    );
+
+    const pointDifference = newPoints - userPoints;
+    await updatePoints(pointDifference);
 
     try {
       await updateDoc(doc(firestore, "tasks", id), {
-        completed: updatedTasks.find((task) => task.id === id)?.completed,
+        completed: updatedTask.completed,
       });
     } catch (error) {
       console.error("Error updating document: ", error);
@@ -71,25 +94,31 @@ export const useTasks = (date: string, user: User) => {
     newTaskDate: string,
     newColor: string
   ) => {
-    const updatedTasks = tasks.map((task) =>
-      task.id === id
-        ? {
-            ...task,
-            text: newTask,
-            description: newDescription,
-            goalId: newGoalId,
-            date: newTaskDate,
-            color: newColor,
-          }
-        : task
-    );
+    const originalTask = tasks.find((task) => task.id === id);
 
-    const updatedTask = updatedTasks.find((task) => task.id === id);
-
-    if (!updatedTask) {
+    if (!originalTask) {
       console.error(`Task with id ${id} not found.`);
       return;
     }
+
+    // Determine if a goalId has been added or removed
+    let pointsChange = 0;
+    if (!originalTask.goalId && newGoalId) {
+      pointsChange += 1; // A goalId was added
+    } else if (originalTask.goalId && !newGoalId) {
+      pointsChange -= 1; // A goalId was removed
+    }
+
+    await updatePoints(pointsChange);
+
+    const updatedTask = {
+      ...originalTask,
+      text: newTask,
+      description: newDescription,
+      goalId: newGoalId,
+      date: newTaskDate,
+      color: newColor,
+    };
 
     try {
       const taskDocRef = doc(firestore, "tasks", id);
@@ -100,15 +129,37 @@ export const useTasks = (date: string, user: User) => {
         date: updatedTask.date,
         color: updatedTask.color,
       });
-      // setTasks(updatedTasks);
     } catch (error) {
       console.error("Error updating document: ", error);
     }
   };
 
   const handleDeleteTask = async (id: string) => {
+    const taskToDelete = tasks.find((task) => task.id === id);
+
+    if (!taskToDelete) {
+      console.error(`Task with id ${id} not found.`);
+      return;
+    }
+
+    let pointsToDeduct = 0;
+
+    // If the task was completed, deduct points
+    if (taskToDelete.completed) {
+      pointsToDeduct += 2;
+
+      // If the task is linked to a goal, deduct additional points
+      if (taskToDelete.goalId) {
+        pointsToDeduct += 1;
+      }
+    }
+
+    await updatePoints(-pointsToDeduct);
+
+    // Filter out the task from the local state
     setTasks(tasks.filter((task) => task.id !== id));
 
+    // Delete the task from Firebase
     try {
       await deleteDoc(doc(firestore, "tasks", id));
     } catch (error) {
@@ -117,8 +168,6 @@ export const useTasks = (date: string, user: User) => {
   };
 
   useEffect(() => {
-    console.log("Fetching tasks for date:", date, "and user:", user.uid);
-
     const taskCollection = collection(firestore, "tasks");
     const q = query(
       taskCollection,
@@ -136,7 +185,6 @@ export const useTasks = (date: string, user: User) => {
           tasksForDay.push(task);
         });
         setTasks(tasksForDay);
-        console.log(tasks);
       },
 
       (error) => {
@@ -155,6 +203,7 @@ export const useTasks = (date: string, user: User) => {
     handleCompleteTask,
     handleEditTask,
     handleDeleteTask,
+    userPoints,
   };
 };
 export default useTasks;
